requirements.txt
streamlit
sentence-transformers
faiss-cpu
rank_bm25
transformers
torch
numpy
nltk
beautifulsoup4
requests
